requirements.txt
gunicorn
flask
google-cloud-dialogflow
python-dotenv
forex_python
google-cloud-translate
requests